Add files via upload
Creates MYSQL db tables, and parses a GEDCOM to populate individuals, families, and children. GEDCOMs are wicked, so dates are text fields to allow for approximate dates and circas, and individuals missing from the GEDCOM but that are still part of the tree are given the individuals.id of '0000' Ditto for children with no families in the tree. The missing family will have an id of '0000'. Hopefully that'll make it easier to find erroneous or missing data later.